consolidate sinatra instrumentation
